refactor: update ticket query in crm.gql
This commit updates the ticket query in crm.gql to include additional fields such as assignee, completedAt, and suspendedAt. This will provide more comprehensive information about the ticket.

Also, update package.json and package-lock.json

This commit adds the "fslightbox-react" package as a dependency in package.json and package-lock.json. This package will be used to implement a lightbox feature in the application.